Code formatting

+++ sources/include/Application.h
@@ -43,8 +43,8 @@ class Application final : public WindowObserver, public Ogre::LogListener, publi
   int64_t time_of_last_frame_ = 0;
   int64_t cumulated_time_ = 0;
   int64_t fps_counter_ = 0;
-  int current_fps_ = 0;
-  int target_fps_ = 60;
+  int64_t current_fps_ = 0;
+  int64_t target_fps_ = 60;
   bool lock_fps_ = true;
   std::string log_;
   std::string log_file_ = "Launch.log";
